moved Socket_Singleton.py into the src/Socket_Singleton dir
