+++ pyproject.toml
@@ -1,7 +1,34 @@
+[project]
+authors = [
+  {name = "Emboiko", email = "ed@emboiko.com"},
+]
+classifiers = [
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.7",
+  "Programming Language :: Python :: 3.8",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Operating System :: OS Independent",
+]
+dependencies = []
+description = "Allow a single instance of a Python application to run at once"
+license = "MIT"
+license-files = ["LICEN[CS]E.*"]
+name = "Socket_Singleton"
+readme = "README.md"
+requires-python = ">=3.7"
+version = "2.0.1"
+
+[project.urls]
+Repository = "https://github.com/emboiko/Socket_Singleton"
+
+[build-system]
+build-backend = "uv_build"
+requires = ["uv_build>=0.11.7,<0.12.0"]
+
 [tool.black]
-line-length = 100
-target-version = ['py37', 'py38', 'py39', 'py310', 'py311', 'py312']
-include = '\.pyi?$'
 extend-exclude = '''
 /(
   # directories
@@ -18,10 +45,12 @@ extend-exclude = '''
   | __pycache__
 )/
 '''
+include = '\.pyi?$'
+line-length = 100
+target-version = ['py37', 'py38', 'py39', 'py310', 'py311', 'py312']
 
 [tool.isort]
-profile = "black"
-line_length = 100
 known_first_party = ["Socket_Singleton"]
+line_length = 100
+profile = "black"
 skip_glob = ["*/__pycache__/*", "*/build/*", "*/dist/*"]
-
