+++ src/Socket_Singleton/__init__.py
@@ -1,0 +1,6 @@
+from .Socket_Singleton import Socket_Singleton, MultipleSingletonsError
+
+__all__ = [
+    "Socket_Singleton",
+    "MultipleSingletonsError",
+]
